be/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1
python-socketio==5.11.0
google-cloud-storage>=2.12.0
celery>=5.4.0
redis>=5.0.0
pymongo>=4.6.0
psutil>=7.0.0
pymilvus>=2.3.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.13
tiktoken>=0.5.0
openai>=1.0.0
